Moved FSM memory functions to separate top-level file compute_finite_state_machine_memory
